feat(ci): maximise strictness of all quality gates
Workspace lints: deny unsafe_code, unused_imports, dead_code,
trivial_casts, trivial_numeric_casts, unused_must_use,
unused_variables, unreachable_patterns. Warn for
missing_debug/copy_implementations.

Clippy pedantic/nursery/cargo lints enabled at warn level with
common noisy ones allowed. unwrap_used, expect_used,
indexing_slicing, string_slice set to warn (tighten to deny
incrementally).

CI: hash-pin all GitHub Actions, add cargo doc check with
RUSTDOCFLAGS="-D warnings", remove blanket RUSTFLAGS="-D warnings"
(rely on workspace lints for per-lint severity), enable SHA pinning
requirement on the repo.

Pre-push hook updated to match clippy invocation (no -D warnings).

.clippy.toml added for cognitive-complexity-threshold.

Fix all newly-denied lints: #[allow(unsafe_code)] on libc calls,
remove trivial usize casts in tests, coerce instead of cast in
backend-local, suppress unused test helper.

+++ crates/cascade/tests/p2p_e2e.rs
@@ -122,8 +122,8 @@ async fn e2e_multiple_files_independent_blocks() {
         .unwrap();
     let bridge = P2pBridge::new(engine, db);
 
-    let data_a = vec![0x11; BLOCK_128KB as usize + 100];
-    let data_b = vec![0x22; BLOCK_128KB as usize * 2 + 200];
+    let data_a = vec![0x11; BLOCK_128KB + 100];
+    let data_b = vec![0x22; BLOCK_128KB * 2 + 200];
 
     let id_a = ItemId::new("p2p", "file_a.bin");
     let id_b = ItemId::new("p2p", "file_b.bin");
@@ -158,7 +158,7 @@ async fn e2e_multiple_files_independent_blocks() {
 /// Build test data spanning several blocks (~384 KB).
 fn build_test_data() -> Vec<u8> {
     // Three full blocks + a partial.
-    let size = BLOCK_128KB as usize * 3 + 7890;
+    let size = BLOCK_128KB * 3 + 7890;
     let mut data = Vec::with_capacity(size);
     for i in 0..size {
         data.push((i % 251) as u8); // PRIME_MOD to get varied byte values.
